pyproject.toml
[project]
name = "mbt-workspace"
version = "0.1.0"
description = "Model Build Tool - A dbt-like framework for ML pipelines"
requires-python = ">=3.10"
dependencies = [
    "mbt-core",
    "mbt-sklearn",
    "mbt-h2o",
    "mbt-mlflow",
    "mbt-postgres",
    "mbt-s3",
    "mbt-airflow",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.uv.workspace]
members = [
    "mbt-core",
    "mbt-sklearn",
    "mbt-h2o",
    "mbt-mlflow",
    "mbt-postgres",
    "mbt-s3",
    "mbt-airflow",
    "examples/telecom-churn",
]

[tool.uv.sources]
mbt-core = { workspace = true }
mbt-sklearn = { workspace = true }
mbt-h2o = { workspace = true }
mbt-mlflow = { workspace = true }
mbt-postgres = { workspace = true }
mbt-s3 = { workspace = true }
mbt-airflow = { workspace = true }
